SoundSystem: fixed stream sounds playback (FINAL)
Tested with OGG and WAV files, with and without repeat.
It's for sure now: it works. Check MantisBT issue #207 for details.

In this version, all basic functionality is working properly.
That includes: {global,individual}_{play,pause,stop,restart}

Individual fade in/out is NOT ready.
It performs badly, though it generates no system crashes.
Global fade in/out has not been tested yet.