SoundSystem: updates for MantisBT issue #392, still no compilation!
Finished SoundHandler API for single sounds playback.
We have to implement everything now, and the current big question is
what to use as SoundManager unique ID.
One possibility is to change the interface and make the SoundManager
generate the IDs of the EnvSounds we make it play. That ID would be
stored in a field in the corresponding SoundHandler structure.

IMPORTANT TODO: change the use of the Playlists memory address
as unique ID for the EnvSounds played from playlists.