SoundSystem: implementation of SoundHandler::SingleSound
The class is fully implemented.
We have also started with the SoundHandler methods that manipulate them.
Currently we're stuck at SoundHandler::startSound()

Next step is making the SoundManager generate its owns play indices
when SoundManager::startEnvSound() is invoked, so we can store those inside
the SingleSound structure in SoundHandler::startSound()