Sound System: updating the SoundManager environmental sounds
To make the SoundManager generate its own ids on every environmental sound
creation, we modified the whole interface of the class regarding (only)
these kind of sounds. The interface is completed, not yet the implementation.

The header (SoundManager.h) contains comments with full details
for the expected behaviour of every method.
The source (SoundManager.cpp) contains a "TODO: continue from here"
in the exact point we left, and there are some ideas commented there
on how to implement this environmental sound id generation.